Added order functionality to front end web app.

+++ ThMoCo.WebApp/Models/UserProfileViewModel.cs
@@ -9,7 +9,7 @@
         public int Id { get; set; }
         public string? Email { get; set; }
         public string? PhoneNumber { get; set; }
-        public double? Fund { get; set; }
+        public decimal? Fund { get; set; }
         public DateTime? UpdatedAt { get; set; }
         public DateTime? LastLogin { get; set; }
         public string? Provider { get; set; } // e.g., "auth0", "google", "facebook"
